requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
redis==5.0.1  # Optional, for Redis caching
